Add files via upload
Added a level and xp system to the code. Now after logging in you can play the famous random game where you have to guess the number between 1 and 10. If you get it correct, you gain an xp point. If you get 10 xp points you level up. All of this information is stored in a mysql database.